# Make various core math functions easier for the compiler to reason about