Details are in the commit messages, but the overall goal here is for a post-#43852 compiler to be able to figure out that all of these are `:effect_free` and `:terminates` as well as `:nothrow` where appropriate. This is achieved by moving the code around a bit and avoiding inference loops. The primary benefit of this is not seen until #43852 is merged, but the change is independent and will be marginally better on the current compiler as well.