@oscardssmith 2022-01-24
Can you post the performance differences?