@Keno 2022-01-24
It lets the compiler reason about them and delete them if unused (and nothrow), e.g.:

master:
```
julia> function foo(x)
           x ^ 2.0
           return x
       end
foo (generic function with 1 method)

julia> @code_typed foo(1.0)
CodeInfo(
1 ─     invoke Main.:^(x::Float64, 2.0::Float64)::Any
└──     return x
) => Float64

julia> using BenchmarkTools
[ Info: Precompiling BenchmarkTools [6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf]

julia> @benchmark foo(1.0)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  1.697 ns … 17.979 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.705 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.777 ns ±  0.506 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▂█                                  ▆▃                 ▁▆ ▁
  ▄██▅▁▃▇▆▁▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃██▁▁▁▆▅▁▁▁▁▁▁▁▁▁▁▁▁██ █
  1.7 ns       Histogram: log(frequency) by time      1.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

PR + #43899:
```
julia> function foo(x)
           x ^ 2.0
           return x
       end
foo (generic function with 1 method)

julia> @code_typed foo(1.0)
CodeInfo(
1 ─     return x
) => Float64

julia> using BenchmarkTools
[ Info: Precompiling BenchmarkTools [6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf]

julia> @benchmark foo(1.0)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  0.023 ns … 11.227 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     0.025 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   0.027 ns ±  0.112 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆  █  ▇  ▇  ▇   ▅  ▁  ▁  ▂         ▁               ▂  ▁    ▂
  █▁▁█▁▁█▁▁█▁▁█▁▁▁█▁▁█▁▁█▁▁█▁▁█▁▁▁█▁▁█▁▁█▁▁▄▁▁▁▁▁▁▇▁▁█▁▁█▁▁▄ █
  0.023 ns     Histogram: log(frequency) by time    0.041 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```