@Keno 2022-01-24
Note that this is quite fancy, because the non-throwness is constant dependent:
```
julia> function foo(x)
           x ^ 2.1
           return x
       end
foo (generic function with 1 method)

julia> @code_typed foo(1.0)
CodeInfo(
1 ─     invoke Main.:^(x::Float64, 2.1::Float64)::Any
└──     return x
) => Float64
```
Because 
```
julia> foo(-1.0)
ERROR: DomainError with -1.0:
Exponentiation yielding a complex result requires a complex argument.
Replace x^y with (x+0im)^y, Complex(x)^y, or similar.
Stacktrace:
 [1] throw_exp_domainerror(x::Float64)
   @ Base.Math ./math.jl:37
 [2] ^(x::Float64, y::Float64)
   @ Base.Math ./math.jl:983
 [3] foo(x::Float64)
   @ Main ./REPL[5]:2
 [4] top-level scope
   @ REPL[7]:1
```