@oscardssmith 2022-01-24
Why is `foo` getting inferred as `Any`?