@Keno 2022-01-24
> Why is `foo` getting inferred as `Any`?

Bit of a display artifact. The type isn't stored if the statement is unused.